converted javascript files to typescript files. initial types implementation. added user model.
