+++ src/components/Logo/Logo.tsx
@@ -1,0 +1,21 @@
+import { Tilt } from "react-tilt";
+import vision from "../../images/vision.png";
+import "./Logo.css";
+
+const Logo = () => {
+  return (
+    <div className="ma4 mt0">
+      <Tilt
+        className="Tilt br2 shadow-2"
+        options={{ max: 55 }}
+        style={{ height: 150, width: 150 }}
+      >
+        <div className="Tilt-inner">
+          <img style={{ paddingTop: "12px" }} alt="logo" src={vision} />
+        </div>
+      </Tilt>
+    </div>
+  );
+};
+
+export default Logo;
