--- src/components/Navigation/Navigation.js
@@ -1,35 +1,0 @@
-import React from "react";
-import ProfileIcon from "../Profile/ProfileIcon";
-
-const Navigation = ({ onRouteChange, isSignedIn, toggleModal, user }) => {
-  if (isSignedIn) {
-    return (
-      <nav style={{ display: "flex", justifyContent: "flex-end" }}>
-        <ProfileIcon
-          onRouteChange={onRouteChange}
-          toggleModal={toggleModal}
-          user={user}
-        />
-      </nav>
-    );
-  } else {
-    return (
-      <nav style={{ display: "flex", justifyContent: "flex-end" }}>
-        <p
-          onClick={() => onRouteChange("signin")}
-          className="f3 link dim white-90 underline pa3 pointer"
-        >
-          Sign In
-        </p>
-        <p
-          onClick={() => onRouteChange("register")}
-          className="f3 link dim white-90 underline pa3 pointer"
-        >
-          Register
-        </p>
-      </nav>
-    );
-  }
-};
-
-export default Navigation;
