+++ src/components/Profile/ProfileIcon.tsx
@@ -1,0 +1,79 @@
+import React from "react";
+import {
+  Dropdown,
+  DropdownToggle,
+  DropdownMenu,
+  DropdownItem,
+} from "reactstrap";
+
+import User from "../../models/User";
+
+interface Props {
+  user: User;
+  toggleModal: () => void;
+  onRouteChange: (route: string) => void;
+}
+
+interface State {
+  dropdownOpen: boolean;
+}
+
+class ProfileIcon extends React.Component<Props, State> {
+  constructor(props: Props) {
+    super(props);
+    this.state = {
+      dropdownOpen: false,
+    };
+  }
+
+  toggle = () => {
+    this.setState((prevState: State) => ({
+      dropdownOpen: !prevState.dropdownOpen,
+    }));
+  };
+
+  render() {
+    return (
+      <div className="pa4 tc">
+        <Dropdown isOpen={this.state.dropdownOpen} toggle={this.toggle}>
+          <DropdownToggle
+            tag="span"
+            data-toggle="dropdown"
+            aria-expanded={this.state.dropdownOpen}
+          >
+            {this.props.user.avatar ? (
+              <img
+                src={this.props.user.avatar}
+                className="br-100 ba h3 w3 dib"
+                alt="avatar"
+              />
+            ) : (
+              <img
+                src={`https://robohash.org/set_set3/${this.props.user.name}`}
+                className="br-100 ba h3 w3 dib"
+                alt="avatar"
+              />
+            )}
+          </DropdownToggle>
+          <DropdownMenu
+            right
+            className="b--transparent shadow-5"
+            style={{
+              marginTop: "20px",
+              backgroundColor: "rgba(255, 255, 255, 0.5)",
+            }}
+          >
+            <DropdownItem onClick={this.props.toggleModal}>
+              View Profile
+            </DropdownItem>
+            <DropdownItem onClick={() => this.props.onRouteChange("signout")}>
+              Sign Out
+            </DropdownItem>
+          </DropdownMenu>
+        </Dropdown>
+      </div>
+    );
+  }
+}
+
+export default ProfileIcon;
