+++ src/components/Rank/Rank.tsx
@@ -1,0 +1,57 @@
+import React from "react";
+
+interface Props {
+  name: string;
+  entries: number;
+}
+
+interface State {
+  emoji: string;
+}
+
+class Rank extends React.Component<Props, State> {
+  constructor(props: Props) {
+    super(props);
+
+    this.state = {
+      emoji: "",
+    };
+  }
+
+  componentDidMount() {
+    this.generateEmoji(this.props.entries);
+  }
+
+  componentDidUpdate(prevProps: Props, prevState: State) {
+    if (
+      prevProps.entries === this.props.entries &&
+      prevProps.name === this.props.name
+    ) {
+      return null;
+    }
+    this.generateEmoji(this.props.entries);
+  }
+
+  generateEmoji = (entries: number) => {
+    fetch(
+      `https://igmf9hvng9.execute-api.eu-central-1.amazonaws.com/dev/rank?rank=${entries}`
+    )
+      .then((response) => response.json())
+      .then((data) => this.setState({ emoji: data.input }))
+      .catch(console.log);
+  };
+
+  render() {
+    return (
+      <React.Fragment>
+        <div className="white f3">
+          {`${this.props.name}, your current entry count is...`}
+        </div>
+        <div className="white f1">{this.props.entries}</div>
+        <div className="white f3">{`Rank Badge: ${this.state.emoji}`}</div>
+      </React.Fragment>
+    );
+  }
+}
+
+export default Rank;
