+++ src/models/User.ts
@@ -1,0 +1,8 @@
+export default interface User {
+  id: string;
+  name: string;
+  avatar?: string;
+  email: string;
+  entries: number;
+  joined: string;
+}
